refactor: 패키지 분류
