+++ src/main/java/finalmission/woowabowling/member/MemberRepository.java
@@ -1,4 +1,4 @@
-package finalmission.woowabowling;
+package finalmission.woowabowling.member;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
